DockerHub/aggregator_node/requirements.txt
paramiko==3.5.0
watchdog==6.0.0
python-dotenv==1.0.1
